The DenseGillespieAlgorithm Module

Module Index

DenseGillespieAlgorithm.chooseevent

DenseGillespieAlgorithm.mainiteration!

DenseGillespieAlgorithm.mainiteration!

DenseGillespieAlgorithm.nexteventandtime

DenseGillespieAlgorithm.onestep!

DenseGillespieAlgorithm.sumsumdict

Detailed API

DenseGillespieAlgorithm.chooseevent — Method

chooseevent(rates::Vector{Float64},total_rate::Float64;<keyword arguments>)

From the vector of total rates choose at random one of the indices of the vector according to their rates. The value 0 is returned if the total rates are positive, but too smale to let the evolution continue. The maximum number of tries is set by max_try=1000.

DenseGillespieAlgorithm.mainiteration! — Method

mainiteration!

For models where population states are dictionaries (or vectors) with the traits as
keys and the subpopulation size as values. Here the population history is
itselfe a dictionary with the same keys and the individual subpopulation
history as a vector for every trait.

DenseGillespieAlgorithm.mainiteration! — Method

mainiteration!

For OneType model where the population state is a number and
the population history is a vector.

DenseGillespieAlgorithm.nexteventandtime — Method

nexteventandtime(rates::Vector{Float64})

Samples a exponential distributed random variable to determine the time for the next event and calls choose_event. The return value is a tuple consiting of the envent index returned by choose_event and the time to the next event.

DenseGillespieAlgorithm.onestep! — Method

Executes one step of the evolution by modifying x_0 and rates.

DenseGillespieAlgorithm.sumsumdict — Method

For a dictionary with vectors of values calculates the sum of all values of all vectors combined.