Function Index

Module Index

DenseGillespieAlgorithm.chooseevent

DenseGillespieAlgorithm.chooseevent

DenseGillespieAlgorithm.dropzeros!

DenseGillespieAlgorithm.dropzeros!

DenseGillespieAlgorithm.historylength

DenseGillespieAlgorithm.mainiteration!

DenseGillespieAlgorithm.mainiteration!

DenseGillespieAlgorithm.nexteventandtime

DenseGillespieAlgorithm.nexteventandtime

DenseGillespieAlgorithm.onestep!

DenseGillespieAlgorithm.run_gillespie!

DenseGillespieAlgorithm.run_gillespie!

DenseGillespieAlgorithm.saveonestep!

DenseGillespieAlgorithm.stop!

DenseGillespieAlgorithm.sumsumdict

Detailed API

DenseGillespieAlgorithm.chooseevent — Method

chooseevent(rates::Vector{Float64},total_rate::Float64)

Choose from the vector of total rates at random one of the indices of the vector according to their rates.
The value 0 is returned if the total rates are positive, but too smale to let the evolution continue.

DenseGillespieAlgorithm.chooseevent — Method

chooseevent(rates::Dict,total_rate::Float64)

Choose from the dictionary of total rates at random one of the keys of the dictionary according to their values.
The value 0 is returned if the total rates are positive, but too smale to let the evolution continue.

DenseGillespieAlgorithm.dropzeros! — Method

dropzeros!(ps)

Do nothing for non-dictionary inputs.

DenseGillespieAlgorithm.dropzeros! — Method

dropzeros!(ps::Dict{Any,Vector})

Eliminates all key value pairs for which the the firts entry of the vector of the value is zero.

DenseGillespieAlgorithm.historylength — Method

historylength(population_hisotry,par)

Return the simulation time based on the length of the population history.
If the population history is neither a `Vector` nor a `Matrix` it is assumed that the Parameter has a field called `historylength` that is then returned.

DenseGillespieAlgorithm.mainiteration! — Method

mainiteration!(pop_hist,rates,n0,ct,time,par,ex!::F1,r!::F2,stat!::F3,hstart)

Mainiteration of the GillespieAlgorithm for complex models.

DenseGillespieAlgorithm.mainiteration! — Method

mainiteration!(pop_hist,rates,n0::Real,ct,time,par,ex!::F1,r!::F2,stat!::F3,hstart)

Mainiteration of the GillespieAlgorithm for OneType Models where the population state is a number and the population history is a vector.

DenseGillespieAlgorithm.nexteventandtime — Method

nexteventandtime(rates::Vector{Float64})

Sample a exponential distributed random variable to determine the time for the next event and calls `choose_event`.
The return value is a tuple consiting of the envent index returned by `choose_event` and the time to the next event.

DenseGillespieAlgorithm.nexteventandtime — Method

nexteventandtime(rates::Dict)

Sample a exponential distributed random variable to determine the time for the next event and calls `choose_event`.
The return value is a triple consiting of the envent index and trait returned by `choose_event` and the time to the next event.

DenseGillespieAlgorithm.onestep! — Method

onestep!(x_0,rates,t_0,t_end,par,ex!::F1,r!::F2)

Execute one step of the evolution by modifying `x_0` and `rates` and returning the current time `t_0`.

DenseGillespieAlgorithm.run_gillespie! — Method

    run_gillespie!(time,n₀,par,execute!,rates!,initrates,population_history[,hstart=0,statistic!])

Run a exact stochastic simulation, return and fill the population_history.

#Arguments

time::AbstracVector: time interval for the simulation

n₀: initial population state

par: additional parameter (gets passed to execute! and rates!)

execute!: execute function

rates!: rates function

initrates: initial rates

population_history: empty population history

hstart=0: time shift for parameter change (opitonal)

statistic!: additional statistic function (optional)

Extended help

Note that n₀,initrates,population_history all three get modified during the simulation

The algorithm expects the execute! function to have the following signature julia execute!(i::Number,n₀,par) where the i is the event that gets executed and the population state n₀ gets modified accordingly. The only exception is when the initrates are given as a dictionary. In that case the signature is execute!(i,trait,n₀,initrates,par), where 'trait' is the key that is modified.

The algorithm expects the rates! function to have the following signature julia rates!(initrates,n₀,par) where the rates get modified according to the current population state given in n₀.

The algorithm expects the statistic! function to have the following signature julia statistic!(population_hist,t,n₀,par) where the population history gets modified at position t with the current population state n₀.

Note that the population_history needs to be accessable via index from 1 to length(time), or if hstart is given from 1+hstart to length(time)+hstart. Unless a specified statistic! function is given.

Note that the initial population state n₀ must match the population_history in the sense that population_history :: Vector{typeof(n₀)}. Unless a specified statistic! function is given.

The parameter variable par is passed through all functions (execute!,rates!,statistics!), thereby affording the user additional flexibility.

DenseGillespieAlgorithm.saveonestep! — Method

 saveonestep!(pop_hist,index,ps,par)

 Sav one step of the simulation. Generic method if no explicit statistic! function is given.

DenseGillespieAlgorithm.stop! — Method

stop!(pop_hist,index,n0,par,stat!)

Fill the remaining population history with the (statistic of) the current population state if the evolution came to a halt.

DenseGillespieAlgorithm.sumsumdict — Method

sumsumdict(D::Dict{String,Vector})

Calculate the sum of the sums of the vectors that are the values of a dictionary.